Move configs under dashboard and add user restrictions
This change
* Moves configs under the Admin namespace and gives the path /admin/configs
* Adds a link for configs in the Admin sidebar
* Adds user restrictions to prevent unauthorized users from accessing configs
