+++ app/views/admin/configs/_config.html.erb
