+++ app/views/admin/configs/_config.json.jbuilder
