+++ app/views/admin/configs/_core_form.html.erb
