+++ app/views/admin/configs/_form.html.erb
