+++ app/views/admin/configs/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @configs, partial: 'admin/configs/config', as: :config
