+++ app/views/admin/configs/new.html.erb
