+++ app/views/admin/configs/new_solr_host.html.erb
