+++ app/views/admin/configs/show.json.jbuilder
