--- app/views/configs/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @configs, partial: 'configs/config', as: :config
