+++ spec/views/admin/configs/_core_form.html.erb_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe 'configs/_core_form' do
+RSpec.describe 'admin/configs/_core_form' do
   context 'when in host setup step' do
     before do
       @config = Config.new(setup_step: 'host')
